MLK-18822 imx: Fix bootaux issue when running on ARM64
The bootaux from community uses ulong to read private data and write to M4 TCM,
this cause problem on ARM64 platform where the ulong is 8bytes.
Fix it by using u32 to replace ulong.

(cherry picked from commit f3d936c84759fcd47a2489cf31fe46cd84ba1f47)

+++ arch/arm/mach-imx/imx_bootaux.c
@@ -13,7 +13,7 @@
 #ifndef CONFIG_IMX8
 int arch_auxiliary_core_up(u32 core_id, ulong boot_private_data)
 {
-	ulong stack, pc;
+	u32 stack, pc;
 
 	if (!boot_private_data)
 		return -EINVAL;
